moved db out of __init__ into database to decrease circular imports
