+++ app/__init__.py
@@ -7,16 +7,16 @@ App builder. Can be imported and used to start the site
 from flask import Flask
 from flask.ext.bootstrap import Bootstrap
 from flask_security import Security
-from flask.ext.sqlalchemy import SQLAlchemy
 from flask_debugtoolbar import DebugToolbarExtension
 import logging
 from raven.contrib.flask import Sentry
 from werkzeug.contrib.fixers import ProxyFix
+
 from config import config
+from .database import db
 
 bootstrap = Bootstrap()
 security = Security()
-db = SQLAlchemy()
 toolbar = DebugToolbarExtension()
 
 from .models import user_datastore
@@ -27,6 +27,7 @@ def create_app(config_name):
     app.config.from_object(config[config_name])
 
     bootstrap.init_app(app)
+
     db.init_app(app)
     security.init_app(app, user_datastore)
     toolbar.init_app(app)
