+++ app/database.py
@@ -1,0 +1,7 @@
+"""
+Database builder
+"""
+
+from flask.ext.sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
